LAB: primes
User pipe and fork to generate a list of prime number.
Initially, fork and put 2 to 35 into a pipe from parent
side, child try to read first item in pipe as prime number.
After that, open a new pipe and fork. In parent side, consume
items from old pipe and check if it is not factor of current
prime number then send those items which meets the condition.
As child side, repeat the process above.